+++ test/parsing_test/parsing_test.js
@@ -1,11 +1,11 @@
 var chai = require('chai');
 var expect = chai.expect; // we are using the "expect" style of Chai
-var parsing = require('../../libs/parsing');
+var parser = require('../../libs/parser');
 var fs = require('fs')
 var path = require('path');
 
 
-describe('Parsing', function() {
+describe('parser', function() {
   var pantip_html = ''
 
 
@@ -17,18 +17,18 @@ describe('Parsing', function() {
   });
 
   it('should have title', function() {
-    title = parsing.getTitle(pantip_html);
+    title = parser.getTitle(pantip_html);
     expect(title).not.to.be.empty;
     expect(title).to.eql('คู่มือ แฉ แบไต๋ "ซื้อประกัน" อย่างคุ้มค่าและไร้กังวล {ลูกค้าอ่านดี ตัวแทนอ่านได้}')
   });
 
   it('should have post story', function() {
-    story = parsing.getStory(pantip_html);
+    story = parser.getStory(pantip_html);
     expect(story).not.to.be.empty;
   });
 
   it('should have tags', function() {
-    tags = parsing.getTags(pantip_html);
+    tags = parser.getTags(pantip_html);
   });
 
 });
